requirements.txt
torch
torchvision
torchaudio
torch-scatter
torch-sparse
torch-cluster
torch-geometric
rdkit
biopython
transformers
sentencepiece
scikit-learn
pandas
matplotlib
seaborn
tqdm
deepchem
captum
PyYAML
streamlit
requests